// migrations/snapshots/1ab9cfe52a1b68fa58ffc1418c732782e9482504299977753047f75f864f60bd/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly pontos: CodecTypes['pg/int4@1']['output'];
      readonly telefone: CodecTypes['pg/text@1']['output'] | null;
      readonly usuarioId: CodecTypes['pg/text@1']['output'];
    };
    readonly Estoque: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly produtoId: CodecTypes['pg/text@1']['output'];
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly unidadeId: CodecTypes['pg/text@1']['output'];
    };
    readonly HistoricoStatusPedido: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly observacao: CodecTypes['pg/text@1']['output'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['output'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
    };
    readonly ItemPedido: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly pedidoId: CodecTypes['pg/text@1']['output'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['output'];
      readonly produtoId: CodecTypes['pg/text@1']['output'];
      readonly quantidade: CodecTypes['pg/int4@1']['output'];
      readonly subtotal: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Pagamento: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly mensagem: CodecTypes['pg/text@1']['output'] | null;
      readonly metodo: CodecTypes['pg/text@1']['output'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['output'];
      readonly status: 'PENDENTE' | 'APROVADO' | 'NEGADO';
      readonly transacaoId: CodecTypes['pg/text@1']['output'] | null;
      readonly valor: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Pedido: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly canalPedido: 'APP' | 'TOTEM' | 'BALCAO' | 'PICKUP' | 'WEB';
      readonly clienteId: CodecTypes['pg/text@1']['output'] | null;
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
      readonly unidadeId: CodecTypes['pg/text@1']['output'];
      readonly valorTotal: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Produto: {
      readonly ativo: CodecTypes['pg/bool@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly preco: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Unidade: {
      readonly ativo: CodecTypes['pg/bool@1']['output'];
      readonly endereco: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
    };
    readonly Usuario: {
      readonly ativo: CodecTypes['pg/bool@1']['output'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly perfil: CodecTypes['pg/text@1']['output'];
      readonly senhaHash: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly pontos: CodecTypes['pg/int4@1']['input'];
      readonly telefone: CodecTypes['pg/text@1']['input'] | null;
      readonly usuarioId: CodecTypes['pg/text@1']['input'];
    };
    readonly Estoque: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly produtoId: CodecTypes['pg/text@1']['input'];
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly unidadeId: CodecTypes['pg/text@1']['input'];
    };
    readonly HistoricoStatusPedido: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly observacao: CodecTypes['pg/text@1']['input'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
    };
    readonly ItemPedido: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['input'];
      readonly produtoId: CodecTypes['pg/text@1']['input'];
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly subtotal: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Pagamento: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly mensagem: CodecTypes['pg/text@1']['input'] | null;
      readonly metodo: CodecTypes['pg/text@1']['input'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly status: 'PENDENTE' | 'APROVADO' | 'NEGADO';
      readonly transacaoId: CodecTypes['pg/text@1']['input'] | null;
      readonly valor: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Pedido: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly canalPedido: 'APP' | 'TOTEM' | 'BALCAO' | 'PICKUP' | 'WEB';
      readonly clienteId: CodecTypes['pg/text@1']['input'] | null;
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
      readonly unidadeId: CodecTypes['pg/text@1']['input'];
      readonly valorTotal: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Produto: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly preco: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Unidade: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly endereco: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
    };
    readonly Usuario: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly perfil: CodecTypes['pg/text@1']['input'];
      readonly senhaHash: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly pontos: CodecTypes['pg/int4@1']['input'];
      readonly telefone: CodecTypes['pg/text@1']['input'] | null;
      readonly usuarioId: CodecTypes['pg/text@1']['input'];
    };
    readonly Estoque: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly produtoId: CodecTypes['pg/text@1']['input'];
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly unidadeId: CodecTypes['pg/text@1']['input'];
    };
    readonly HistoricoStatusPedido: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly observacao: CodecTypes['pg/text@1']['input'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
    };
    readonly ItemPedido: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly precoUnitario: CodecTypes['pg/numeric@1']['input'];
      readonly produtoId: CodecTypes['pg/text@1']['input'];
      readonly quantidade: CodecTypes['pg/int4@1']['input'];
      readonly subtotal: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Pagamento: {
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly mensagem: CodecTypes['pg/text@1']['input'] | null;
      readonly metodo: CodecTypes['pg/text@1']['input'] | null;
      readonly pedidoId: CodecTypes['pg/text@1']['input'];
      readonly status: 'PENDENTE' | 'APROVADO' | 'NEGADO';
      readonly transacaoId: CodecTypes['pg/text@1']['input'] | null;
      readonly valor: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Pedido: {
      readonly atualizadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly canalPedido: 'APP' | 'TOTEM' | 'BALCAO' | 'PICKUP' | 'WEB';
      readonly clienteId: CodecTypes['pg/text@1']['input'] | null;
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly status:
        | 'CRIADO'
        | 'AGUARDANDO_PAGAMENTO'
        | 'PAGO'
        | 'EM_PREPARACAO'
        | 'PRONTO'
        | 'ENTREGUE'
        | 'CANCELADO';
      readonly unidadeId: CodecTypes['pg/text@1']['input'];
      readonly valorTotal: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Produto: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly preco: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Unidade: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly endereco: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
    };
    readonly Usuario: {
      readonly ativo: CodecTypes['pg/bool@1']['input'];
      readonly criadoEm: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly perfil: CodecTypes['pg/text@1']['input'];
      readonly senhaHash: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;